+++ launchDaemonManagement.zsh
@@ -24,6 +24,7 @@
 #
 # Version 2.4.0b2, 02-Feb-2026, Dan K. Snelson (@dan-snelson)
 # - Added space-delimited list of `acceptableAssertionApplicationNames` (Feature Request #67; thanks for the suggestion, @yassermkh!)
+# - Added Dark Mode Overlay Icon [Feature Request #62](https://github.com/dan-snelson/DDM-OS-Reminder/issues/62) (thanks for the suggestion, @cyberotterpup!)
 #
 ####################################################################################################
 
@@ -38,7 +39,7 @@
 export PATH=/usr/bin:/bin:/usr/sbin:/sbin:/usr/local:/usr/local/bin
 
 # Script Version
-scriptVersion="2.3.1b1"
+scriptVersion="2.4.0b1"
 
 # Client-side Log
 scriptLog="/var/log/org.churchofjesuschrist.log"